+++ tests/contrib/multi_tenancy/test_tenant_models_using.py
@@ -52,15 +52,25 @@ class Product(TenantModel):
         is_tenant = True
 
 
+class Cart(TenantModel):
+    products = fields.ManyToMany(Product)
+
+    class Meta:
+        registry = models
+        is_tenant = True
+
+
 @pytest.mark.parametrize("use_copy", ["false", "instant", "after"])
 async def test_schema_with_using_in_different_place(use_copy):
     if use_copy == "instant":
         copied = models.__copy__()
         NewTenant = copied.get_model("Tenant")
         NewProduct = copied.get_model("Product")
+        NewCart = copied.get_model("Cart")
     else:
         NewTenant = Tenant
         NewProduct = Product
+        NewCart = Cart
     tenant = await NewTenant.query.create(
         schema_name="edgy", domain_url="https://edgy.dymmond.com", tenant_name="edgy"
     )
@@ -68,10 +78,25 @@ async def test_schema_with_using_in_different_place(use_copy):
         copied = models.__copy__()
         NewTenant = copied.get_model("Tenant")
         NewProduct = copied.get_model("Product")
+        NewCart = copied.get_model("Cart")
+    cart = await NewCart.query.using(schema=tenant.schema_name).create()
+    assert cart.__using_schema__ == tenant.schema_name
     for i in range(5):
-        await NewProduct.query.using(schema=tenant.schema_name).create(name=f"product-{i}")
+        product = await NewProduct.query.using(schema=tenant.schema_name).create(
+            name=f"product-{i}"
+        )
+        if i % 2 == 0:
+            product_through = cart.products.through(cart=cart, product=product)
+            product_through.__using_schema__ = tenant.schema_name
+            assert await cart.products.add(product_through)
+        else:
+            assert await cart.products.add(product)
 
     total = await NewProduct.query.filter().using(schema=tenant.schema_name).all()
+
+    assert len(total) == 5
+
+    total = await cart.products.filter().using(schema=tenant.schema_name).all()
 
     assert len(total) == 5